make commit for tabular version
this version allows there to be a tab after the initial
line number, instead of what used to be 2 spaces.
i found that when i was trying to import this into the siemens system, it was giving me a lot of werrors with
spaces.  the tabs work fine, and it is the way the code
downloads anyway