Removed on_change arg from show_input_panel
When the on_done method is called as the adjust_line_start and
adjust_line_increment are being edited, it triggers a bug in the
renumbering code. So I removed the call to self.on_done, from the
on_change arg of show_input_panel, and put it in the on_done position
only.